meli/lib/requirements.txt
requests
firebase
firebase_admin
schedule
pillow
